Corrected a stupid amount of wrong data types (This will help lower memory usage. Untested, but did compile)

+++ src/server/game/AI/CoreAI/UnitAI.cpp
@@ -196,7 +196,7 @@ void UnitAI::FillAISpellInfo()
             UPDATE_TARGET(AITARGET_SELF)
         else
         {
-            for (uint32 j = 0; j < MAX_SPELL_EFFECTS; ++j)
+            for (uint8 j = 0; j < MAX_SPELL_EFFECTS; ++j)
             {
                 uint32 targetType = spellInfo->Effects[j].TargetA.GetTarget();
