+++ src/server/game/Handlers/CharacterHandler.cpp
@@ -1471,7 +1471,7 @@ void WorldSession::HandleEquipmentSetSave(WorldPacket &recv_data)
     eqSet.IconName  = iconName;
     eqSet.state     = EQUIPMENT_SET_NEW;
 
-    for (uint32 i = 0; i < EQUIPMENT_SLOT_END; ++i)
+    for (uint8 i = 0; i < EQUIPMENT_SLOT_END; ++i)
     {
         uint64 itemGuid;
         recv_data.readPackGUID(itemGuid);
@@ -1507,7 +1507,7 @@ void WorldSession::HandleEquipmentSetUse(WorldPacket &recv_data)
 
     sLog->outDebug(LOG_FILTER_NETWORKIO, "CMSG_EQUIPMENT_SET_USE");
 
-    for (uint32 i = 0; i < EQUIPMENT_SLOT_END; ++i)
+    for (uint8 i = 0; i < EQUIPMENT_SLOT_END; ++i)
     {
         uint64 itemGuid;
         recv_data.readPackGUID(itemGuid);
